requirements.txt
nltk==3.8.1
python-docx==1.1.0
PyPDF2==3.0.1